New Templates

+++ src/main/java/br/com/linx/cardValidator/model/BinType.java
@@ -22,7 +22,7 @@ public class BinType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_type")
-    private int idType;
+    private Long idType;
 
     private String description;
